[Template merge - langs/und] Added support for alternate writing systems for spellers.
[Template merge - langs/und]
Fixed a build bug for MacVoikko, causing the final target to always be out of date. Regulated verbosity for zhfst targets. The twolc testing scripts now print a message when there is no test data. The Hfst twolc testing script properly detects when there is no test data, and exits with the SKIP (77) value.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/evn@111896 c7155fb1-f0a7-4240-a2fc-2600b6f42f90